Parses options partially (correct length, text, etc.)
When a content change is passed to parsing an option partially, parser
can handle the change and return a partially updated result, or null if
it wants to bail out.

A lot of code in this commit is duplicated from OptionParser and/or can
be generalized, which I plan to do after doulbe-checking the concepts
implemented here in the headline parser.