Fenced code blocks work
Fenced code blocks (with triple backticks) work as they were in the old
browser. Info-blocks are parsed into the options (the first word of the
info block is the default option, the rest is "codeBlockInfo"). When
there is an explicit options block, there cannot be an info block.

Not implemented from the github spec: Indentation of the fences and the
block (1-3 spaces are allowed according to GFM)

https://github.github.com/gfm/#code-fence