Refactored how we know all parsers
There is now an interface "Parsers" and implementations of that
interface hold information of all known parsers for a certain type of
document.

The only productive implementation is "MfMParsers" (for Marmota flavored
Markdown). If one of the parsers needs another parser to do its work
(like OptionsParser needs OptionParser), it must query the
implementation of the Parsers interface that it got during construction.

This refactoring enabled me to remove a FIXME: A fenced code block can
now decide whether it was ended by another element by querying the two
element parsers from the Parsers implementation it knows.