Refactor: Removed ParseResult intermediate object
Before this change, all parsers returned a ParseResult intermediate
object, which was no longer necessary. The original purpose of this
object was to provide the start index and the length of some content.

But now, all content that can be returned from parsers is of type
Updatable<T> and thus contains this information already. Because of
that, the intermediate object was not necessary anymore, and I removed
it.