New, documented design; Parsing single option works
There are now some design documents in the "docs" folder that describe
how parsing and parial parsing is supposed to work and the data
structures that support this process.

`Option` and `OptionParser` already use this new data structure, and all
the tests pass. I also added some tests for `TextParser`, so that
partial parsing is not only tested in the elements that use it, but also
tested in a more general way.

Everything else is still quite a mess, with compiler errors and skipped
tests. There are even a lot of renamed test files because the tests
won't even start - there are too many errors.