Partially parsing options works for first line
Partially parsing multi-line options is a bit tricky and can only be
fully implemented once the code of the parser is cleaned up. But now, it
already works for the first line, and we have a (failing, skipped) test
in place for the second line.

Also, for now, this only works from the POV of the `OptionsParser`. The
`TextParser` itself still has a too simplistic implementation for
partially parsing content - It always parses the whole line, even when
there are parts of the line that could be partially parsed
independently.