OptionsParser: Cleanup and bug fixes
Cleaned up the `OptionsParser` so that there is only one version of the
code that parses the options.

This lead to a small design change in the text parser, where parsing a
single piece of content now only returns that content, not the
containing element, too. That also required a few changes in how we test
`OptionParser`.

Then, it became possible to fix a few defects in `OptionsParser`:
- Partially parsing multi-line options works
- Default options are prevented from the second line on
- When an options block is completely parsed, one cannot add another
  line