Partially parses only parts of an element, if possible
When an `UpdatableElement` (eg. `UpdatableOptions`) contains multiple
other elements, then the parital parser will first try to only partially
parse those elements. Only when that fails, it will partially parse the
complete element.

Consider the following options block:

{ default-value; key1=value1, key2    = a long   value; }

If a content change fits inside one of the `Option` elements, e.g.
removing the spaces between "long" and "value" inside the value of
"key2", then the text parser will first attempt to only partially parse
this `Option` element.

If that succeeds, we are done. If it does not (returning `[null,
null]`), the parser will attempt to parse the complete options block. If
that fails (e.g. because someone added a closing bracket), it will
attemt to parse the parent of that options block, and so on.