Main parser passes single lines to sub-parsers
The main parser `Marmdown` parses the document line-by-line and passes
each line to all the parsers. If one parser returned a result for a
line, `Marmdown` will also try to parse the **next** line with that
parser first. So, that parser will get fed consecutive lines as long as
it can parse the previous one.

Only then, when the previous parser could not parse a line, `Marmdown`
tries to parse it with all other parsers again.

We do this to ensure that multiline content is parsed by the same parser
and that no higher-level parser can interrupt it.