Ends block on empty lines (denoted by \r and/or \n)
Lines can now end in either \r or \n or \r\n.

When there are one or more empty lines, `Marmdown` does not pass the
next line to the previous parser, thus ending the block. This meand that
we cannot have multi-line text content that also contains empty lines,
and that is **on purpose**.