Can parse multi-line heading with multi-line options
The heading parser now parses multi-line headings considering multi-line
options: If there is an options block in the heading that is not
finished yet, the heading parser first passes the current line to the
options parser.

Because of that, the options parser now has a chance to finish the
options block. Only then, the heading parser tries to parse the
remaining line as heading text.