Can partially parse headings
Partially parsing headings faced two major problems:
* The lenght of ParsedDocumentContent instances was fixed and not
  updated when partially parsing
* The heading level was fixed at creation and not automatically computed
  from the heading's content

Now, the length of (almost) all ParsedDocumentContent instances is
calculated based on their contents (except for StringContent which has a
fixed length).

The heading level is computed from the heading's content: The heading
finds the StringContent that represents the start of the heading (e.g.
'###') and takes the length of that content, which **is** the heading
level.