Added JSDoc documentation infrastructure
Creating the documentation for Marmdown can now be done with

```
npm run doc
```

That creates the API documentation in the output folder "out". Also, all
the markdown files from "./docs" are used as "tutorials" (the
JSDoc-feature the project is using to add markdown files). Those
so-called "tutorials" can be linked from the API documentation (and
vice-versa).