Minimal set of element types for an MfM document
The minimal set of types for the MfM dialect is now defined. BUT this
is NOT the minimal set of elements to parse valid MfM documents - It
is just the minimal set of elements so that the dialect compiles
(Container, Section, Paragraph and Text).

With this set of elements, it will now be possible to implement a first
proof-of-concept parser that can ONLY parse text-only paragraphs.

The different element types (ContainerBlock, LeafBlock, ContainerInline
and LeafInline) are all interfaces now, which makes implementing the
types for a dialect easier. The helper classes for implementing the
concrete element types are now in GenericElement.ts.

Since elements need unique IDs, there is now an IdGenerator interface
(with a default implementation), and MfMDialect is already using it.