MfMDialect uses LineByLineParser to parse document
The `MfMDialect` uses the `LineByLineParser` in combination with the
`MfMContainerParser` to parse the complete document.

`LineByLineParser` splits the document into lines and then uses the
given container parser to create the component: It passes the current
line **and** the previous result to the container parser, asking it
to update the previous result.

When the line-by-line parser detects an error (either the container
parser returns `null` in the middle of the document or it returns a
different container element in the middle of the document), it throws
a `ParserError`.