Container parser uses section parser
A MfM container consists of document options and one or more sections.

As a first step, the container parser passes a document that consists
of only one line to the section parser, but does nothing with the
result yet.

With this test, we now have the infrastructure for managing all parsers
of a dialect in place: The `Parsers` interface collects all known
parsers of a dialect. The class `MfMParsers` is the only concrete
implementation so far, and it contains all parsers that form the
MfM dialect.

The class `MfMDialect` has exactly one object of this class. Since
parser implementations are stateless, this object can be shared across
all parsers and all documents of a dialect.