Can parse `MfMContainer` elements
The parser for `MfMContainer` can now parse all simple cases (no
options, not parsing updates).

The code might even be reusable: It just generically goes over a list
of sub-parsers and uses the first one that can parse the element. When
there is a previous element, it tries to extend that first.