MfMHeading returns the meta block MfMSection
When a heading is parsed, it will always be wrapped in a section. So,
the heading now returns a meta block. To do that, I had to define
how meta blocks are being handled (meta-blocks.md in the docs folder)
and extend the parser interface to allow returning of meta blocks.

Also, the container parser now ignores the previous element if it has
the "fully parsed" property set to true.