Parses the text of the heading using TextParser
The heading parser knows a parser for the heading text content, which
in turn uses "all leaf parsers" to parse the heading content. In the
test, only the text parser is used.

Also, there is still a problem with type safety here - "all leaf
parsers" should return a type better than "Element", so we can then
make sure that it returns one of the leaf elements.