Started integration tests for Github-flavored-Markdown
In ATXHeading.spec.ts, there is a first integration test for the
parts of the MfM Dialect that are compatible with GfM, and the test
case comes directly from the GfM specification:
https://github.github.com/gfm/#example-32

The implementation is not yet complete, though: It does not convert
the markdown to HTML correctly. But the test would not run yet
anyway - The parser for the dialect does not invoke the heading
parser yet.