MfMContainer always contains sections
MfM Documents are entirely and exclusively composed from sections.
To achieve that, the parser for MfM Container creates a section as
previous element when there is no previous container yet.

Also, MfMContainerParser now uses the `allBlocks` property of it's
`_parsers` property to loop over all possible blocks, to make the
code more generic. In practice, it will always only parse headings
and sections, because of the way headings and sections work.