Sections and headings are nested correctly.
Both MfMContainer and MfMSection now use a general "parseBlock" function
to parse its content. This parseBlock function makes sure that sections
are nested correctly.

Because of that, the result for the first "ATXHeading" test almost looks
complete already: Only the text of the heading is still missing. To
make it pass, MfMTextParser must be implemented first to parse the
unformatted heading text.