Readability improvement: Nesting of LineContent
Before this change, LineContent could be deeply nested because every
instance of LineContent had a content array (even leaves like
StringLineContent).

Now, the data structure states more clearly how content can be nested:
There is ParsedLine which represents a "real" line of the document (or
a part of a single line) and it implements LineContent. This class can
contain other instances of LineContent.