All currently implemented elements remember content lines
The elements that are currently implemented all remember line content
correctly: They keep track of an array of ParsedLine objects that represent
a single line **or** a part of a single line that belongs to the current
element.

Those ParsedLine objects, in turn, contain all the parts that make up the
content of the current element. They can, e.g., be of type StringLineContent
or be a ParsedLine of an inner element or any other implementation of
LineContent.

Since the complete text of a document can be reproduced by combining the
text of all it's lines, the library can now reproduce the content of a
very simple document (see AsText.spec.ts).