Add integration tests for document structure
The structure of a parsed markdown document is important for rendering
a presentation. And it cannot be fully tested with `html(...)`, since
there are invisible elements in a structure (like sections).

In order to be able to test this, there are now some integration tests
that test the document structure based on a string that can be created
from this document structure.