UpdateParser changes ID of updated element
The update parser now changes the ID of the updated element, since this
element does not have the original content anymore. In most cases, this
will hopefully be enough, but there will be cases where it also needs
to update the ID of one or more container elements; e.g. when an option
changes, the whole options block logically has changed and also the
Element which contains the options must change.

Since element IDs can be used to trigger re-renders in the UI and an
option change on one element might even cause other elements to change,
it might become even more complicated.

But I'll solve **that** problem when I get to it.