Updating multi-line headings works
The `UpdateParser` can now update multi-line elements, TDDed through
multi-line headings. It also updates the start indexes of everything
that comes after the updated element, even higher up in the tree (but
that probably needs more thorough testing).

Multi-line headings can now reproduce the original text of each line;
That was not possible before because the heading did not put the store
spaces at the end of the line in its `lines` content.