Refactor: Abstract MfMBlockContentParser
MAKE THE CHANGE EASY: MfMGeneralPurposeBlockParser does not implement
the parser directly anymore. It now uses a more generic MfMBlockContentParser
that can be configured to accept any tokent to start a line with.

There is, right now, only this one extension to the more generic parser:
The MfMGeneralPurposeBlockParser which uses '>' as the token and
parses block quotes. BUT with this change, implementing the parser for
asides (which use the token '^' in MfM) will be really easy.