More tests for nesting blocks, fixed asText defect
While adding more tests for nested blocks, I found a defect when
recreating the original text from a markdown document. Empty lines
were not processed correctly:

* The container parser added too much content from the original document
  in the "empty" line
* It also added the empty line twice, because when it tried to parse
  another line, that **try** already added a line

The code that prevents the double-add mentioned above is not great,
but it works for now. And there is a comment as a reminder of that
refactoring opportunity, for later.