Adds options to headings
When a heading has an options block (indicated by a { immediately after
the heading token, e.g. "#{ ...the options... } Heading test), the
heading parser parses those options and adds them to the heading.
Otherwise it adds empty options-so, a heading always has options.

The heading parser also makes sure that, when there are options that
are not fully parsed (e.g. because there is no closing } in the second
line of the options block or because the second line of the options
block is syntactically incorrect), the heading cannot be updated during
update parsing.

The above would be a problem when one makes the second options line
in the example below syntactically correct by inserting an = sign between
"key2" and "value2":

```
    #{default option
    key2 value2 } Heading Text
```

In that case, we must parse the complete heading again, because the
second line must change from simple text content to options+text content.