Added (and disabled) legacy tests from marmota.app
The legacy tests shall later make sure that we support the same use
cases as the legacy parser from marmota.app, even though this parser
is not a drop-in replacement. They need some work to make them run again
and that rewriting should make sure that they match the original intent.