Asides and General Purpose Blocks have Options
Both Asides and General Purpose Blocks can have options now. The tests
for this feature are in the MfMGeneralPurposeBlock test file, since
Asides work exactly the same.

The code here needs quite a bit of refactoring: Adding options to some
content should be a cross cutting concern-It should be *really* easy.
But right now, the implementations for headings and for the gp-Blocks
are completely different.