Refactor: parseBlock does not create element anymore
The helper function parseBlock, that MfMContainer and MfMSection use
to parse their content, does not create the element anymore. It gets
the previous container and the container to use as parameters instead.

That change made some implicit kowledge in MfMContainer obsolete. And
it will it possible later to add options to MfMContainer, since the
parser can now parse the options before handing a new container off
to parseBlock.