MfMContainer can have options
The MfMContainer can now have options at the beginning of a document.

Refactor: There is now a base class for MFM Elements that handles all
the common stuff, like that elements have options or when elements are
fully parsed.

Fix: There was some ambiguity when a new parsed line had to be pushed
to an element, which did not work in all cases, especially for options.
Now, the MfMOptions parser always calls a callback `addLine` when adding
options to an element, and the caller can decide whether adding a line
is necessary or not.

Fix: MfMContainer cannot process updates directly, only their content
can be updated.