Fixed changing a paragraph line into a different element
When a line of a paragraph was changed by an `update` so that it
had to be parsed as a different element, the result of parsing the
update was wrong. e.g. when an update would insert "# " at the beginning
of a paragraph line, the update parser would have still parsed this
line as a paragrpah and added the "# " as text.

Now, the update parser **always** bails out when an update to an element
could change the document structure, e.g. when it adds text at the
beginning of a paragraph line. Better bail out too often than not often
enough...

Also, the update parser cannot create a default id generator anymore.
Because that was possible before and was used incorrectly, there were
situations where duplicate IDs could have been supplied.