Refactor: Container blocks with dynamic lines (2)
Sections now use dynamic lines. Since sections and containers were parsed
by the same funciton, I had to temporarily duplicate this function.

Also, a container with dynamically created lines now throws an error
when some code is trying to push a line (instead of printing a warning).
This makes it easier to find those places in the code.