Refactor: Container blocks with dynamic lines (3)
MfMContainer now uses dynamic lines to create it's line content.

Attaching information to a line is now more generic: It can not only
attach content to prepend or append to a line, it can also attach other
information (e.g. whether the line was fully parsed).

Deciding whether the last options line and the first content line should
be merged-which can happend when there is content in the same line as
the last options line-is now done based on attached information (see
above).

When updating an element, all information about the updated lines is now
re-attached to any container that might contain updated lines. Because
of that, the line content of container elements can be recreated correctly
after updates.