MfMEmphasis can parse updates and options.
Updates and options can now be parsed correctly. Options of inline
elements can only contain a single line.

Also, emphasis that contain an unclosed left / right delimiter run are
now parsed in a way so that the outer emphasis is parsed as emphasis,
not as text span. To make that possible, inline parsers now accept
an "additional parameters" - object; The emphasis parser uses this
object to pass a function that can determine whether an outer emphasis
should be ended by a delimiter run.