Can parse indented code blocks
Since MfM has indented code blocks now, parsing updates must be done
very carefully at the start of a line: When an update changes the
indentation, it might change the document structure! e.g. it might change
a simple paragraph line to a code block...

Because of that, the update parser now refuses all updates that might
change the indentation of a block.