Links are compatible with the legacy parser.
The legacy parser tests did assume that the options for images contain
a "video" type when the destination could be a video. I did not want
to bake this functionality right into the parser, so I added "options
postprocessors" instead: When creating the MfMDialect, one can support
"MfMDialectOptions" that can contain options postprocessors for certain
elements. Now, adding the "video" type is only done within the legacy
tests, using an options postprocessor.