Fixed problem when updating dynamic line content
When some content was updated inside a block with dynamic line content,
the update did not work. The updated line now has the correct "belongs-to"
element, and that allows the element with the dynamic lines to calculate
the lines correctly again.

Also, line IDs now have their own type (`line-${string}`) which makes the
code that assigns them more robust.