Lists: Can parse list tasks
A list taks, in marmdown, is the `[ ]` or `[x]` part of a task list
item. So, implementing this was the first step towards implementing
task list items, and thus, generalized list items.

Note that list tasks, in marmdown, can also be `[]`, which is not
allowed in GfM. With this feature, checking a list item by removing
the space and then inserting the 'x' works via updates to the task
list item.