Parses multi-line options on multi-line list items
The parser can now parse multi-line options on multi-line list items.
There were some problems in the code for handling the first line of
a list item that had to be fixed beforehand; e.g. the way it handled
spaces before the "real" item content did not work for more generalized
cases.

Also, the code for handling the content after the end of the options
block has now been factored into its own function. Now, that part of
the list item can be parsed using the function, whether it is on the
first line or any subsequent line.