Parses updates on list items.
Parsing updates in the content and options works for single-line and
multi-line list items. Updating the text within the list item itself
does not work, since doing so would require updating dynamic lines.
That would be hard because the line itself does not really exist - The
parser would have to update the attachments within that dynamic line.